Show the messages in target language of MT context
Trick the RL module fetcher to fetch the module and messages for
the target language in the context. Just change the wgUserLanguage
before calling mw.loader.using and reset just after the module is
fetched.

One issue is the hacky part of modifying wgUserLanguage for moments,
but I believe the purpose of the extension justify it.

Bug: T214010
Change-Id: I971da9b1eba1fa6feb1e0711afa4a3bff4d1613b