Rewrite the menu URLs when source language is not English
A very simple approach that uses the event-name data attributes
set by MF is used to map special page titles to canonical titles.

The langlinks API won't work for special pages. The titles are not
always special pages too. So instead of using sophisticated apis
to find exact title in target language, mapping to English URLs
is done. That is guaranteed to work in any wiki.

Bug: T215241
Change-Id: I4b35571ee38adefad926e3b4a5f8031c62f6dc18