Fix broken MT service info overlay
Clicking on Machine Translation icon was not showing the overlay
with machine translation info, but was causing a js error down
in the stack: 'Uncaught TypeError: e.on is not a function'

Since Id0922af6dbe5726d6b93b11e797295afa783d8d3, overlay manager
expect the overlays as synchronously returned.

This also means, this machine translation info overlay was broken
since Feb 2020.

Bug: T267299
Change-Id: Ic096a19d3e8998490a5e95c61ba6066d9a469878