Update Codex Radio markup
The markup for CSS-only Codex Radios has changed[1] and an upcoming
style change in Codex v2.0.0-rc.1 will cause a vertical misalignment
between the Radio icon and label with the old Radio markup.

This patch updates the Radio markup to match the current spec to
avoid such a visual regression.

[1] https://doc.wikimedia.org/codex/latest/components/demos/radio.html#markup-structure

Bug: T392973
Change-Id: I69138e0687ff6b3cfab230c6943ee9105777a018